[PERF] account_report,account_asset: Avoid useless __get__
Description

When opening a report with a lot of data to display,
getting some values can be very costly.

In this case, for the depreciation schedule, we are
retrieving the column expression and currency symbol
each time if not found in `_build_column_dict`.

To avoid retrieving a value that was already fetched,
we can pass the currency and the expression directly.
This reduces unnecessary `__get__` calls.

Benchmark (made in 17.0)

| N° of assets | Before  | After  |
|--------------|---------|--------|
|          500 |    2 s  | 1.55 s |
|           7k |   20 s  |   19 s |
|          17k |   50 s  |   41 s |

Reference

opw-4287192

closes odoo/enterprise#75208

X-original-commit: b492160bc43e6d90464fb0de9a9a7f16096fd6f1
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Aurélien Marie (auma) <auma@odoo.com>