[FIX] sale_subscription: avoid multiple trigger of SO alert
Steps
-----
- Subscriptions > Configuration > Alerts.
- Create an alert triggered when the stage goes from 'Initial' to
'In Progress', set 'Create next activity', chose Email as activity.
- Create a subscription, confirm it.
** Multiple activities created in the chatter (1 would be expected) **

Cause
-----
The `pre_domain_filter` of the `base.automation` is not set, leading to
an activity being created every time a write happens on the sale order
while it is in the `stage_to_id` stage.

Change
-----
Add a pre-domain/domain filter to only trigger actions when the stage
changes.

opw-4273625

closes odoo/enterprise#74635

X-original-commit: 81e9e45cfe095130aa2faf128ccdab386114b5b4
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>