[FIX] website_sale_subscription: correct combination pricing
Steps
- Have a subscription product with two variants, each having one
different possible pricing.
- Go to the eCommerce page of this product and add to cart the
second variant (the one which is not selected by default).
** The variant is added to the cart with the wrong plan **

Cause
When only 1 pricing is possible for a combination, there's no select
element on the page to select a plan and we fallback on the data attribute
of the `add_to_cart` button.
https://github.com/odoo/enterprise/blob/e9f74bce15e564695733577c181ce3f6d6d063a8/website_sale_subscription/static/src/js/website_sale_subscription.js#L15
However, this element is only rendered once for the initial product combination
and never updated when changing to a new combination.
https://github.com/odoo/enterprise/blob/e9f74bce15e564695733577c181ce3f6d6d063a8/website_sale_subscription/views/templates.xml#L90-L94

opw-4197528

closes odoo/enterprise#74405

X-original-commit: 17bed067d58bcd8b6c4802f85b4b2da6d6753f46
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>