[FIX] account_reports: fix isNextLineChild
Currently, `isNextLineChild` method will consider
`~account.report~14|~res.partner~10|0~account.move.line~32` being a
a child of
`~account.report~14|~res.partner~10|0~account.move.line~3` as the first
string starts with the second one.
This is wrong, they are siblings.

The fix here is to add a pipe, as done in `isLineChildOf`.

task-4321032

closes odoo/enterprise#77764

X-original-commit: c9c4f748cc448887a39c587499fe30495e6a1c9a
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Arnaud Sibille (arsi) <arsi@odoo.com>