[FIX] l10n_ae_corporate_tax_report: undefined comparison
Opening the corporate tax report more than once resulted in an error, making it impossible to open the report.
The reason for this was an undefined comparison filter when opening the report for the second time. This was a result of setting the export mode to `file` when opening the report, which in turn makes the comparison undefined.
This commit fixes this by temporarily setting the export mode to `file` when creating the entry.

closes odoo/enterprise#73678

Ticket: https://www.odoo.com/odoo/project/49/tasks/4288546
X-original-commit: 8877efa20f9af814f5f51a441fd59dc83fc7c4c4
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>