[IMP] l10n_mx_edi: delete imported invoice
When you import an invoice with CFDI, the CFDI fields are populated with the invoice state set to draft.
The current behaviour is to prevent invoice deletion when we have the CFDI.
But when it's an imported invoice, we should allow the user to delete the invoice if needed.

This commit add a new condition on delete invoice, so now an CFDI sent invoice with state to draft can be deleted.

task-4306416

closes odoo/enterprise#73680

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>