[FIX] spreadsheet_edition: display spreadsheet as soon as possible
Before this commit, when a list/pivot/graph object is inserted in a
spreadsheet, it was loaded **and then** the spreadsheet was displayed.
It could cause a confusion for the user because, as soon as he clicks on
insert, he receives a notification that the spreadsheet is created in
Documents and then he has to wait for the spreadsheet to be displayed.

With this commit, the spreadsheet is displayed as soon as possible and
the list/pivot/graph object is loaded in the background.

For the pivot, the spreadsheet is turned into readonly mode until the
pivot is loaded, as we need the pivot to be loaded to know the cells to
insert.

closes odoo/enterprise#74537

Task: 4256694
X-original-commit: b1a344e8b6e90726564311d08988436a36239e0b
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>