[IMP] helpdesk: fix auto assignment on batch ticket create
The auto assignement feature has never worked when creating tickets in batch (all created tickets get assigned to the same user for each team).
The reason was that we were concerned about the impact on performance.

However, it turns out that the impact is really not significant.
With this commit, for a batch of 500 tickets in a team with the "randomly" assigment method,
determining users to assign takes around 10.2 ms, or less than 0.25% of the `create` total execution time.

While there is currently no scenario in standard where we create tickets in batch, it could be useful for customer customizations and avoids having to use loops to create multiple tickets in tests.

Task-4297562

Part-of: odoo/enterprise#74027
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>