[PERF] sale_account_accountant: improve so candidates query plan
The query in `_get_invoice_matching_so_candidates` will frequently
plan a Seq Scan because of the low selectivity of the conditions.
And when it does not, the number of rows returned by the CTE is
expected to high. Because of that, filtering the resulting rows
by a sequence of `OR` conditions can quickly become slow.

In this commit, the `OR` conditions are replaced by `LIKE(ANY(ARRAY[]))`
conditions. That way the CTE is only referenced once in the conditions
and applying the filter on each row is way faster. We also
discarded duplicated `text_tokens` to reduce the size of the ARRAY.

The reason why using a LIKE(ANY(ARRAY[])) is faster is because
postgres inlines the CTE in the outer query. This means that it
performs a Seq Scan on sale.order and injects the CTE definition of
sub.name inside the WHERE conditions of the outer query, along with injecting
the WHERE conditions of the CTE.

So, the regex functions are distributed among the OR conditions. I.e.
every OR condition left operand will contain the regex functions. As those
have to be executed for every OR condition, this quickly becomes slow.

We can explicitely materialize the CTE to avoid that. This makes postgres evaluate
the regex functions only once. But it will still have lots of OR conditions to check
along with running pattern matching for each one of them.

LIKE(ANY(ARRAY[])) avoids this issue. The CTE is still inlined but since we now only have
a single condition, the regex functions are only evaluated once and pattern matched once against
an array of options. This makes the whole query faster and scale better.

speedup

Customer database with 808341 sale.orders. Query timing when increasing
the number of text tokens.

| Number of tokens | Before PR | After PR |
|        2         |     7s    |    1.3s  |
|        5         |     10s   |    1.3s  |
|        10        |     18s   |    1.4s  |
|        20        |     33s   |    1.55s |

opw-4329067
opw-4316765

closes odoo/enterprise#74872

X-original-commit: 749dc487a30cadc2b83da7cff14aa5133cb3dfae
Signed-off-by: William André (wan) <wan@odoo.com>