[FIX] stock_barcode: assign mto moves in barcode split
Steps to reproduce:

- In the settings: enable multi-steps route
- Inventory > Configuration > Warehouse Management > Routes
- Unarchive MTO
- Create a storable prodcut with MTO, buy routes and a set vendor
- Create and confirm a sale order for 3 units of your product
- Confrim the purchase order and the associated reciept
- Go to the barcode app on you delivery
- set the quantity to 1/3 and leave the barcode app
- Go back to the the delivery
> only one line 1/1 appears

Cause of the issue:

The barcode app relies on stock move lines to work properly.
These move lines are even in correspondance with the barcode lines.
When you update a line in the barcode to be 1/3 and then leave the app,
to keep track of the demand, the app split the moves in 2:
https://github.com/odoo/enterprise/blob/69338098f5407a01b4ddcb8b18bc82cf9d7cb4d5/stock_barcode/models/stock_move.py#L9-L11
To create the new barcode lines, it is therefore expected that new move
lines will be created by this call for the splitted move. which is the
purpose of this line:
https://github.com/odoo/enterprise/blob/69338098f5407a01b4ddcb8b18bc82cf9d7cb4d5/stock_barcode/models/stock_move.py#L37
However, this will only work for moves whose procure method is not MTO
because mto moves are not confirmed but flaged as waiting and hence not
assigned by the `_action_confirm`:
https://github.com/odoo/odoo/blob/e7b269206fa125532c9e3ab6ed2b632198199cd8/addons/stock/models/stock_move.py#L1356-L1362
https://github.com/odoo/odoo/blob/e7b269206fa125532c9e3ab6ed2b632198199cd8/addons/stock/models/stock_move.py#L1380
https://github.com/odoo/odoo/blob/e7b269206fa125532c9e3ab6ed2b632198199cd8/addons/stock/models/stock_move.py#L1418-L1422
As a result no move line will be generated for these splitted mto moves
and there will be no stock move line for the barcode to rely on.

opw-4232106

closes odoo/enterprise#75090

X-original-commit: 3e7424101f2cd43943e148bfcf6a426af6ad5b1d
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>