[FIX] use context's language on mail template render
Before this commit, the language of the logged in user and the language given in kwargs would be used to translate the content. The email layout would translate to the user language and the body content would use the kwarg's language. This lead to translation discrepancies. After this commit, only the language given in the kwargs is used and therefore fixing the translations issues.

closes odoo/enterprise#74646

X-original-commit: dbcbf68131571ad629d19e42669e95c1459a4760
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>