[FIX] pos_barcodelookup, pos_hr_barcodelookup: disable product creation for non employees
Currently if employes (without users) log into the register and use the
barode reader, if the barcode isn't found in the database, they will be
prompted with the create product screen

Steps to reproduce:
-------------------
* Connect to the db with Mitchell Admin
* Open the **Point of Sale** App
* Selet **Configuration** > **Settings**
* Select the Furniture shop
* Enable **Log in with Employees**
* Open Furniture shop session
* Select an employee that is not a user as employee (Abigail Peterson for ex)
* Either select the barcode icon and present a barcode to the camera or go
into debug mode and enter the barcode in the debug window and select scan.
Prefer to use a barcode not in the db.
> Observation: A popup appear to create a product

Why the fix:
------------
Currently employees are able to create product if the underlying user who
is connected has the group `base.group_system`.
https://github.com/odoo/enterprise/blob/3bf3400a60436e9f07f23c6ba11e5749e4c00459/pos_barcodelookup/static/src/overrides/components/barcode_reader/barcode_reader.js#L8

Employees do not have the right to create products. This commit
https://github.com/odoo/odoo/commit/e10768b6ce5c82e67e14641233dc9deefb48f8ba
already removed the rights to see the create product button from the menu.
We follow the same logic to block the creation process when scanning barcodes.

We also move the product creation outside the BarcodeReader service.

opw-4255570

closes odoo/enterprise#74067

X-original-commit: f41d2c2a0ef3facac46ef37741e6704c1ff66edc
Related: odoo/odoo#187669
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>