[FIX] hr_payroll_sale_commission: test if pricelist
If a pricelist matches the test product, the commission will be
affected. This is one more repeat of
odoo/odoo@abe8e38ccc18cfba77317d3fdaf4a066b5f69b79: it's possible for
a module to set up a universally applicable pricelist (like
`pos_pricer` which sets a 20% discount on everything). In the presence
of such a module, tests with hard-coded "output" prices will fail.

A common workaround is to disable all pricelists when creating the
test, but in this case it seems reasonably easy (and correct) to use
the SO's total amount as computation basis.

closes odoo/enterprise#73818

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>