[FIX] pos_preparation_display: fix skipChange pdis
In the PoS restaurant if you marked a line as skipChange not to send it
on the preparation display, then add the same product to the order and
send it to the preparation display. At this point everything would be
correct, only one of the product would be one the pdis. But if you add
another different product and send it to the pdis, it would also bring
back the one that was marked as skip change.

Steps to reproduce:
-------------------
* Setup a preparation display for a PoS restaurant
* Open PoS and add product A to the order
* Double click on the order line to mark it as skip change
* Add the same product to the order and click on the order button
* At this point the preparation display should only show one product A
* Add another different product B to the order
* Click on the order button
> Observation: The preparation display shows the product B and a second
product A

Why the fix:
------------
Instead of filtering based on the product we filter directly based on
the line UUID that is unique and should avoid any further errors

opw-4322597

closes odoo/enterprise#73967

X-original-commit: a04e5d9ceca25447ad35d44dd23c583301a343f2
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>