[IMP] helpdesk: make "Tickets Analysis" report clickable
The report is related to the model `helpdesk.ticket.report.analysis`,
but we want the user to be redirected to the list view of `helpdesk.ticket`.

So, deleting the prop `disable_linking` on the views' tag isn't enough.
We had to extend their renderer to override `openView`, and change the
resModel to redirect to the proper model.

Luckily, the search view of the report's action inherits from the search
view of the base action of the tickets, without adding any field that
isn't on the base model, which means the filter set on the former
will be applicable after redirecting to the latter.

related: https://github.com/odoo/odoo/pull/187262
task-4334996

closes odoo/enterprise#73828

Related: odoo/odoo#187262
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>