[IMP] hr_payroll_holidays: Refine time off to defer error
Currently, we check time off to defer over the batch period for all
the employees in the batch if there is a regular payslip.

Now, we only check the employees covered by a regular payslips, this
shouldn't block the end of the year bonus for instance if the regular
payslip in the batch is not related to the bonus employee.

closes odoo/enterprise#73878

Taskid: 4328786
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>