[IMP] whatsapp: Prefixed domainless url buttons with base url
Whenever a WhatsApp template has a url button with website_url that has no domain (e.g., /my/home),
this website_url intuitively is expected to be prefixed with the current database domain (e.g., https://odoo.com/my/home).
This domain prefixing takes effect when the template is submitted for approval, or in the list/form view during its creation.

This feature is useful in case we want to create a template button redirecting to the Odoo database that created it.

task-4241442

Part-of: odoo/enterprise#72116
Related: odoo/upgrade#6777
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>