[FIX] l10n_mx_reports: Restricts reports with incorrect code
Steps to reproduce the issue:

1. Activate the Odoo Mexican Localization Reports module
2. In a Mexican company, create a new Account with 1 as code
3. Go to Trial Balance and download COA SAT (XML)
5. In the General Settings with developer mode active, Download XSD files
6. Go to Trial Balance and download COA SAT (XML) again
7. You get a UserError with an unclear message

Explanation:

The Mexican Chart of Accounts have clear rules regarding `account.account.code`.
The only way to verify those accounts is through the XSD files check, but they are not automatically downloaded and the error received with those files downloaded is not user friendly.

Fix reasoning:

Instead of regulating the code when downloading the XML. We'll add warnings on the Chart of Accounts to notify the user when a code is incorrect.
To make the report error clearer to the user, we added a RedirectWarning that displays the accounts with faulty codes before generating the xml.

opw-4287338

closes odoo/enterprise#81724

X-original-commit: 9c84313c5e179f5fd9a3cc4440d1ddaf113d738a
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Diego Hernández González (dihg) <dihg@odoo.com>