[IMP] web_studio: autofocus field label
Before this commit, when drag/drop a new field in the arch, one had to click on the label to modify it.
Because we assume that modifying the field's label is first and main operation that one would want to do.

In addition to handling this case, we actually implement the autofocus every time we enter a field node.

task-4337512

closes odoo/enterprise#73987

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>