[IMP] hr_payroll: improve employee creation in payslip
This PR enhances the payslip process by restricting the creation of employees
directly from payslips. Creating an employee on the fly during payslip
generation is not practical, as a valid employee contract is a essential for
processing any payslip. Without an active contract, it is neither meaningful nor
functional to create or associate an employee with a payslip.

task-4312449

closes odoo/enterprise#74154

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>