[FIX] sale_amazon_avatax: only show the tax excluded amount
When synchronizing orders from Amazon with Avatax enabled, sales order
lines currently show no taxes and incorrectly set the tax included
price as being the subtotal.

This occurs because the fiscal position used does not include any tax,
which is the expected behavior for Avatax. The recomputation then uses
the full amount of the line (which is tax included), and in the absence
of tax, consider this is also the subtotal.

We now intentionally do not consider the tax amount from Amazon on the
order lines, because the distribution of the Amazon's given tax amount
is unknown. The multiple real taxes across that line is not an
information that Amazon gives us, and needs to be seen from Avatax.

This setup ensures that the subtotal on sales order lines remain
accurate.

opw-4214775

closes odoo/enterprise#75027

X-original-commit: c2b7e145d2db13647283e0e4a8bd55c3465b1f02
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>