[IMP] currency_rate_live: Add exchange rate provider for Hungary - MNB
Objective
---------
Many countries want to use the official rates provided by their National
Bank in their database. A lot of them were already implemented, and now
it's the turn of the Hungarian National Bank.

Solution
---------
The MNB has implemented a SOAP service to access the latest rates.

The service is available at this address
http://www.mnb.hu/arfolyamok.asmx?wsdl

and the service operation `GetCurrentExchangeRates()`

returns:
```xml
<MNBCurrentExchangeRates>
  <Day date="2024-11-20">
    <Rate unit="1" curr="AUD">252,78</Rate>
    <Rate unit="1" curr="BGN">209,49</Rate>
    <Rate unit="1" curr="BRL">67,24</Rate>
    ...
  </Day>
</MNBCurrentExchangeRates>
```

Therefore, we can:
1. Open a Zeep client to communicate with the service
2. Parse the XML and fetch the data (currency, ratio, rates and date).
3. Update the result dict accordingly

task-4320506

closes odoo/enterprise#74664

X-original-commit: 70a6d0db138f03facc487fcb2f56a98f2d7470a5
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>