[FW][FIX] sale_purchase_inter_company: right team for SO
Sales order created from purchase orders thanks to the inter-company module were assigned the default team of the company, regardless of the customer configuration.

This commit makes sure the team is correctly computed in the dedicated compute, ensuring the expected team is assigned to the Sales Order.

opw-4166598

closes odoo/enterprise#73833

Forward-port-of: odoo/enterprise#72767
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>