[FIX] mrp_mps: non-default period changes
- hide Order buttons:
When the date type shown (year, month, week, day) is different from the one
defined in the settings, we hide the order buttons.

- disable max_to_replenish_qty:
`max_to_replenish_qty` should only be applied for the period defined in the settings.
Otherwise we have situation where a limit of 15 per month is applied for a whole year.

- make replenish input cells readonly:
When showing a bigger period type than the one selected in the settings
(e.g. year > week), the replenish input fields are made readonly.

This is because there's a loss of precision when showing bigger period
and changing values may have unintended consequences that way.

Also remove the colors on the replenish cells when in readonly.

Part-of: odoo/enterprise#74051
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>