[FIX] mrp_mps: use real dates with lead times
When a bom has a lead time, we want to use the real date to compute the
indirect demand forecast of the components in the MPS.

Currently when showing the MPS in a period other than 'day', Odoo will
do the sum of all forecast records for each period and base its calculation
on the 1st day of the period.
E.g. for the month of January, it will do the sum of all forecasts of
that month then calculate lead times based on the 1st of January.
> indirect_demand_qty
> > Jan-25: 1000
> > Feb-25: 500

This causes issues with lead times when selecting bigger periods.
This fix ensures that the forecasts are stored per date:
> demand_qty_dict
> > Jan-25
> > > 01-01-25: 200
> > > 01-07-25: 500
> > > 01-23-25: 300
> > Feb-25
> > > ...

That way, when calculating lead times, it uses the real date for each
forecast instead of the 1st day of the period shown.

closes odoo/enterprise#74051

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>